CLN: tidy up references and unit testing classes
