--- pyfibre/io/pyfibre_workflow_io.py
@@ -1,10 +1,0 @@
-from pyfibre.io.utilities import load_json
-
-from pyfibre.pyfibre_runner import PyFibreRunner
-
-
-def load_pyfibre_workflow(filepath):
-
-    data = load_json(filepath)
-
-    return PyFibreRunner(**data)
