+++ pyfibre/tests/probe_classes/analyser.py
@@ -1,9 +1,9 @@
-from pyfibre.model.core.base_analyser import BaseAnalyser
+from pyfibre.model.core.base_multi_image_analyser import BaseMultiImageAnalyser
 
 from .multi_images import ProbeMultiImage
 
 
-class ProbeAnalyser(BaseAnalyser):
+class ProbeAnalyser(BaseMultiImageAnalyser):
 
     def __init__(self, *args, **kwargs):
         kwargs['multi_image'] = ProbeMultiImage()
